refactoring: move other data-variables to one 'data' file, move some functions to special files
