+++ src/pages/sets.js
@@ -1,0 +1,5 @@
+export const pageSets = {
+  onCreate: () => {
+    // const mainPageContainer = create document fragment
+  },
+};
